mailer and reset password token moved to config and services folder
