+++ services/key_generator_service.js
@@ -1,0 +1,9 @@
+// Crypto module is part of Node
+const crypto = require("crypto");
+
+// Generates a random key
+function generateRandomKey() {
+    return crypto.randomBytes(20).toString("hex");
+}
+
+module.exports = generateRandomKey;
